18 | Added Soft Delete in Project Feature

+++ CVBuilder.Application/Features/Projects/Commands/DeleteProject/DeleteProjectCommandHandler.cs
@@ -24,6 +24,16 @@ namespace CVBuilder.Application.Features.Projects.Commands.DeleteProject
 
             if (projectToDelete == null)
                 throw new Exceptions.NotFoundException(nameof(Project), request.ProjectId);
+            
+            if (request.SoftDelete)
+            {
+                projectToDelete.IsDeleted = true;
+
+                await repository.UpdateAsync(projectToDelete);
+
+                return Unit.Value;
+            }
+
 
             await repository.DeleteAsync(projectToDelete);
